Update to be node 24 compatible; bump nan and node-gyp dependency version (#374)
* node 24 compat; nan dep bump

* use Nan::Persistent for back compat; add macro helper for property callback back compat; microtasks back compat

* update deasync dep ver to fix CI for Windows; add node 24 to CI matrix; add python-setuptools to macOS CI script; restore flag that were wrongly removed in previous commit

* ci: set versions to 20,22,24

* doc: uppdate supported versions

* tests: make them pass

* Update gst plugin path from brew for macos CI

* Return value for index getter in cairo

* Fix debug printf GType format on Windows x64

* ci: run (empty commit)

* ci: use ubuntu-latest (24.04)

* refactor: Nan::Intercepted

* chore: gitignore

* refactor: more Nan::Intercepted

* Revert a change made by mistake in an earlier commit in PR #374

* Update libsoup to 3.0 for Ubuntu CI

* Update union and struct fields tests to check for init values

* Skip union fields test and some others until bug is resolved

---------

+++ src/debug.cc
@@ -10,6 +10,7 @@
 #include "type.h"
 
 #include <cstdio>
+#include <cinttypes>
 
 namespace GNodeJS {
 
@@ -245,16 +246,16 @@ void print_gtype (GType type) {
 void print_type (GType type) {
     GType t = type;
     while(!G_TYPE_IS_FUNDAMENTAL(type) && t != 0) {
-        printf("\t\x1b[92m -> %s \t (%lu)\x1b[0m\n", g_type_name(t), t);
+        printf("\t\x1b[92m -> %s \t (%" PRIuPTR ")\x1b[0m\n", g_type_name(t), (uintptr_t)t);
         t = g_type_parent(t);
     }
 }
 
 void print_klass (void * klass) {
     printf("\x1b[38;5;202m");
-    printf("klass: %s  (type: %lu )",
+    printf("klass: %s  (type: %" PRIuPTR " )",
             G_OBJECT_CLASS_NAME (klass),
-            G_OBJECT_CLASS_TYPE (klass) );
+            (uintptr_t)G_OBJECT_CLASS_TYPE (klass) );
     printf("\x1b[0m\n");
 }
